refactor pipeline schedule modules
